Avoid Cart Store dependency by moving handleRenewNowClick to separate module
