+++ client/blocks/credit-card-form/index.jsx
@@ -19,7 +19,8 @@ import FormButton from 'components/forms/form-button';
 import formState from 'lib/form-state';
 import notices from 'notices';
 import { validatePaymentDetails } from 'lib/checkout';
-import { handleRenewNowClick, isRenewable } from 'lib/purchases';
+import { handleRenewNowClick } from 'lib/purchases/renew-now';
+import { isRenewable } from 'lib/purchases';
 import ValidationErrorList from 'notices/validation-error-list';
 import wpcomFactory from 'lib/wp';
 import { AUTO_RENEWAL, MANAGE_PURCHASES } from 'lib/url/support';
