+++ client/lib/purchases/index.js
@@ -6,14 +6,11 @@
 
 import { find, includes } from 'lodash';
 import moment from 'moment';
-import page from 'page';
 import i18n from 'i18n-calypso';
 
 /**
  * Internal dependencies
  */
-import analytics from 'lib/analytics';
-import { getRenewalItemFromProduct } from 'lib/cart-values/cart-items';
 import {
 	isDomainRegistration,
 	isDomainTransfer,
@@ -22,7 +19,6 @@ import {
 	isTheme,
 	isConciergeSession,
 } from 'lib/products-values';
-import { addItems } from 'lib/upgrades/actions';
 
 function getIncludedDomain( purchase ) {
 	return purchase.includedDomain;
@@ -73,28 +69,6 @@ function getName( purchase ) {
 
 function getSubscriptionEndDate( purchase ) {
 	return purchase.expiryMoment.format( 'LL' );
-}
-
-/**
- * Adds a purchase renewal to the cart and redirects to checkout.
- *
- * @param {Object} purchase - the purchase to be renewed
- * @param {string} siteSlug - the site slug to renew the purchase for
- */
-function handleRenewNowClick( purchase, siteSlug ) {
-	const renewItem = getRenewalItemFromProduct( purchase, {
-		domain: purchase.meta,
-	} );
-	const renewItems = [ renewItem ];
-
-	// Track the renew now submit.
-	analytics.tracks.recordEvent( 'calypso_purchases_renew_now_click', {
-		product_slug: purchase.productSlug,
-	} );
-
-	addItems( renewItems );
-
-	page( '/checkout/' + siteSlug );
 }
 
 function hasIncludedDomain( purchase ) {
@@ -408,7 +382,6 @@ export {
 	getPurchasesBySite,
 	getRenewalPrice,
 	getSubscriptionEndDate,
-	handleRenewNowClick,
 	hasIncludedDomain,
 	isCancelable,
 	isPaidWithCreditCard,
